refactor: unite weighted sampling to 1 module, change name to SimilarityWeightSampler
