--- simba/core/data/sampling/__init__.py
@@ -1,1 +1,0 @@
-"""Data sampling utilities for weighted sampling in training."""
