+++ simba/core/data/weighted_sampling.py
@@ -1,25 +1,19 @@
 import numpy as np
-
 from simba.utils.binning import round_to_ordinal
+import numpy as np
+import torch
+from torch.utils.data import WeightedRandomSampler
 
 
-class WeightSampling:
+class SimilarityWeightSampler:
     """
-    functions to execute weighted sampling for tackling unbalance in similarities
+    Class for computing weights for weighted random sampling based on binned similarity values.
     """
 
     @staticmethod
     def compute_weights(binned_list):
         freq = np.array([len(r) for r in binned_list])
 
-        # remove 1 from using the last bin for sim=1
-        # index_half = int((len(binned_list)) / 2)
-
-        # for the ranges that are lower than 0.5 treat them as an only range
-        # sum_freq_low_range=sum(freq[0:index_half])
-        # freq_low_range= sum_freq_low_range/(len(freq[0:index_half]))
-
-        # freq[0:index_half] = freq_low_range
         weights = np.zeros(len(binned_list))
         for index, f in enumerate(freq):
             if f != 0:
@@ -27,21 +21,10 @@ class WeightSampling:
             else:
                 weights[index] = 0
 
-        # deprecated
-        # for the ranges that are lower than 0.5, put half the weight on the highest range
-        # weights[0:index_half] = 0.5 * weights[0:index_half]
-
-        # for the weights below 0.5 increase the wiehgt to reduce false positives:
-        # weights[0:index_half] = 2 * weights[0:index_half]
-
         # normalize the weights
         weights = weights / np.sum(weights)
-        # weights= [1 for w in weights]
 
-        # the last bin corresponds to sim=1. So if there are 6 bins, actually there are 5 bins between 0 and 1
-        # bin_size = 1 / (len(binned_list) - 1)
-
-        # currently, we do not bin sim=1
+        # compute the range of similarity values for each bin
         bin_size = 1 / len(binned_list)
         range_weights = np.arange(0, len(binned_list)) * bin_size
         return weights, range_weights
@@ -55,18 +38,6 @@ class WeightSampling:
         bin_size = 1 / (len(binned_list) - 1)
         range_weights = np.arange(0, len(binned_list)) * bin_size
         return weights, range_weights
-
-    # @staticmethod
-    # def compute_sample_weights(molecule_pairs, weights):
-
-    # get similarities
-    #    sim = molecule_pairs.pair_distances[:, 2]
-    #    # sim = [m.similarity for m in molecule_pairs]
-    #    #index = [math.floor(s * (len(weights) - 1)) for s in sim]
-    #    index = [math.floor(s * (len(weights))) if s != 1 else (len(weights)-1) for s in sim  ]
-    #    weights_sample = np.array([weights[ind] for ind in index])
-    #    weights_sample = weights_sample / (sum(weights_sample))
-    #    return weights_sample
 
     @staticmethod
     def compute_sample_weights(
@@ -86,8 +57,6 @@ class WeightSampling:
         # Calculate the index using vectorized operations
         effective_range = len(weights) - 1 if bining_sim1 else len(weights)
 
-        # print(f'sim: {sim}')
-        # print(f'effective range: {effective_range}')
         indices = np.floor(sim * (effective_range)).astype(int)
 
         # make sure the indexes are not below 0
@@ -98,10 +67,6 @@ class WeightSampling:
 
         # Map the indices to weights and normalize
 
-        # print(f'Weights: {weights}')
-        # print(f'indices: {indices}')
-        # print(f'{weights.shape}')
-        # print(f'{indices.shape}')
         weights_sample = weights[indices]
         if normalize:
             weights_sample /= weights_sample.sum()
@@ -114,7 +79,6 @@ class WeightSampling:
         sim = molecule_pairs.pair_distances[:, 2]
 
         # Calculate the index using vectorized operations
-        # indices = np.ceil(sim * (len(weights)-1)).astype(int)
         indices = round_to_ordinal(sim * (len(weights) - 1)).astype(int)
         indices[indices == len(weights)] = len(weights) - 1
 
@@ -123,3 +87,20 @@ class WeightSampling:
         weights_sample /= weights_sample.sum()
 
         return weights_sample
+
+
+class CustomWeightedRandomSampler(WeightedRandomSampler):
+    """WeightedRandomSampler except allows for more than 2^24 samples to be sampled"""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def __iter__(self):
+        rand_tensor = np.random.choice(
+            range(0, len(self.weights)),
+            size=self.num_samples,
+            p=self.weights.numpy() / torch.sum(self.weights).numpy(),
+            replace=self.replacement,
+        )
+        rand_tensor = torch.from_numpy(rand_tensor)
+        return iter(rand_tensor.tolist())
